Schema transport/hr, soft delete, transaksi, log sistem, karyawan, multi-role, dan fitur operasional
Database (supabase/migrations 20260923000003, 20260924000001–000029):
- Pindah semua tabel ke schema transport; schema hr + tabel karyawan
- Soft delete (kolom status 1/2) di semua tabel, transaksi jalankan_transaksi
- Log sistem wajib lengkap (karyawan + IP) dari sesi login; zona waktu WIB
- Unit Trailer + Jenis Unit Trailer (terhubung ke jenis unit), trailer di job
- Menu Karyawan dengan status aktif; driver dan pengguna terhubung ke karyawan
- Satu akun banyak role (role aktif per sesi), edit akun sendiri, aktifkan
  pengguna hanya bila karyawan aktif; tutup celah ubah role oleh non-superadmin
- Ganti truk di job + riwayat; tagihan hanya untuk job tervalidasi
- Slot foto surat_timbang -> surat_jalan; status unit Terjual & Diafkirkan

Backend: modul karyawan, log_sistem, unit_trailer; ganti role & sesi login;
pencarian job berdasarkan no. penawaran; uang jalan tanpa job dibatalkan.

Frontend: halaman Karyawan, Log Sistem, Unit Trailer; pilih/ganti role;
LoadingOverlay, PageHeader, Combobox; link penawaran di daftar job.

Mobile: istilah & slot surat jalan.

+++ backend/app/modules/uang_jalan/schemas.py
@@ -77,14 +77,18 @@ class UangJalanInput(BaseModel):
     job_id: str = Field(min_length=1)
     jenis: UangJalanJenis
     tanggal: str = Field(min_length=1)
-    jumlah: float
+    # Rupiah penuh — kolomnya BIGINT. Sengaja int, bukan float: pecahan yang
+    # lolos ke sini akan dibulatkan diam-diam oleh Postgres, jadi lebih baik
+    # ditolak di pintu masuk. Nilai <= 0 ditangani `_validate` agar pesannya
+    # tetap berbahasa Indonesia.
+    jumlah: int
     sumber_dana_id: str | None = None
     keperluan: str | None = None
     catatan: str | None = None
 
 
 class SetPaguRequest(BaseModel):
-    pagu: float = Field(ge=0)
+    pagu: int = Field(ge=0)
 
 
 class JobUangJalan(BaseModel):
